Finish Day 8
Solve this days problem using python.

### Problem:
Given a grid of nodes find all antinodes.
- Part 1: An antinode is a position that is on the grid and has a distance x to one node and 2x to another node of same type.
- Part 2: An antinode is on a line between 2 nodes of the same type

### Solutions:
For both problems construct a nodes list for each type of node. To find antinodes compare each node of a type with every  other node of that type and add all antinode positions to a set.